Changes Made - Removed setup.sh and replaced with init.sh and setup.mjs. - init.sh is a simple script to install necessary packages and prompt the user to set up the environment using setup.mjs. - setup.mjs is a comprehensive script that sets up the environment for the AgentGPT project. This script interacts with the user to collect necessary configuration details like the OpenAI API key, Docker/Node setup preference, and more. It generates the necessary environment variables, handles Docker setup if chosen, and also has logic to handle and retry on errors.
