+++ .vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "i18n-ally.localesPaths": [
+        "public/locales"
+    ]
+}
