+++ init.sh
@@ -3,9 +3,9 @@
 echo "setting up environment"
 
 # Install inquirer
-#npm init -y
-#npm install inquirer figlet dotenv
+npm init -y
+npm install inquirer dotenv
 echo "done."
 
 echo "Welcome to AgentGPT! The following questions will help you set up your environment"
-node input.mjs
+node setup.mjs
